rf: `buildPropertySchema` (Flow, TS) fns > common fn
Summary:
This is a task of https://github.com/facebook/react-native/issues/34872
- combined `Flow` and `TypeScript` `buildPropertySchema` fn 's into single common
- uses callback param `resolveTypeAnnotation`
- moved it to `parsers-commons.js`
- re-organize imports and exports
