+++ packages/react-native-codegen/src/parsers/typescript/modules/index.js
@@ -22,9 +22,11 @@ import type {
 } from '../../../CodegenSchema.js';
 
 import type {ParserErrorCapturer, TypeDeclarationMap} from '../../utils';
+
 const {nullGuard} = require('../../parsers-utils');
-const {visit, isModuleRegistryCall} = require('../../utils');
+const {visit, isModuleRegistryCall, verifyPlatforms} = require('../../utils');
 const {resolveTypeAnnotation, getTypes} = require('../utils.js');
+
 const {
   unwrapNullable,
   wrapNullable,
@@ -32,8 +34,9 @@ const {
   parseObjectProperty,
   emitUnionTypeAnnotation,
   translateDefault,
-  translateFunctionTypeAnnotation,
+  buildPropertySchema,
 } = require('../../parsers-commons');
+
 const {
   emitBoolean,
   emitDouble,
@@ -50,6 +53,7 @@ const {
   emitMixedTypeAnnotation,
   typeAliasResolution,
 } = require('../../parsers-primitives');
+
 const {
   UnsupportedArrayElementTypeAnnotationParserError,
   UnsupportedGenericParserError,
@@ -57,11 +61,8 @@ const {
   IncorrectModuleRegistryCallArgumentTypeParserError,
 } = require('../../errors.js');
 
-const {verifyPlatforms} = require('../../utils');
-
 const {
   throwIfUntypedModule,
-  throwIfModuleTypeIsUnsupported,
   throwIfUnusedModuleInterfaceParserError,
   throwIfModuleInterfaceNotFound,
   throwIfModuleInterfaceIsMisnamed,
@@ -349,50 +350,6 @@ function translateTypeAnnotation(
   }
 }
 
-function buildPropertySchema(
-  hasteModuleName: string,
-  // TODO(T108222691): Use flow-types for @babel/parser
-  property: $FlowFixMe,
-  types: TypeDeclarationMap,
-  aliasMap: {...NativeModuleAliasMap},
-  tryParse: ParserErrorCapturer,
-  cxxOnly: boolean,
-): NativeModulePropertyShape {
-  let nullable = false;
-  let {key} = property;
-  let value =
-    property.type === 'TSMethodSignature' ? property : property.typeAnnotation;
-
-  const methodName: string = key.name;
-
-  ({nullable, typeAnnotation: value} = resolveTypeAnnotation(value, types));
-  throwIfModuleTypeIsUnsupported(
-    hasteModuleName,
-    property.value,
-    property.key.name,
-    value.type,
-    language,
-  );
-
-  return {
-    name: methodName,
-    optional: Boolean(property.optional),
-    typeAnnotation: wrapNullable(
-      nullable,
-      translateFunctionTypeAnnotation(
-        hasteModuleName,
-        value,
-        types,
-        aliasMap,
-        tryParse,
-        cxxOnly,
-        translateTypeAnnotation,
-        language,
-      ),
-    ),
-  };
-}
-
 function isModuleInterface(node: $FlowFixMe) {
   return (
     node.type === 'TSInterfaceDeclaration' &&
@@ -535,6 +492,9 @@ function buildModuleSchema(
           aliasMap,
           tryParse,
           cxxOnly,
+          language,
+          resolveTypeAnnotation,
+          translateTypeAnnotation,
         ),
       }));
     })
